first step of rework

+++ commands/buy.js
@@ -3,32 +3,29 @@ module.exports = {
 	name: 'buy',
 	summary: 'Buy an item from the shop',
 	description: 'With this you can buy an item from the shop.\nYou can either use `buy <item> <amount> to instantly buy the items or just use `buy`.\nIf you use the latter you will get prompted to enter the name and amount of the item that you want into the chat.',
-	category: 'money',
+	category: 'misc',
 	aliases: ['get'],
 	usage: '<item> <amount>',
 	cooldown: 5,
 	args: false,
 
-	async execute(msg, args, msgUser, profile, guildProfile, bot, options, logger, cooldowns) {
+	async execute(message, args, msgUser, profile, guildProfile, client, logger, cooldowns) {
 
-
-		const avatar = msg.author.displayAvatarURL();
-
-		const filter = m => m.author.id === msg.author.id;
+		const filter = m => m.author.id === message.author.id;
 		let amount = 0;
 		let temp = '';
 		let item;
 
 		const embed = new Discord.MessageEmbed()
-			.setTitle('Neia Shop')
-			.setThumbnail(avatar)
+			.setTitle('DMMO Shop')
+			.setThumbnail(message.author.displayAvatarURL())
 			.setDescription('What item do you want to buy?')
-			.setColor(msgUser.pColour)
+
 			.setTimestamp()
-			.setFooter('Neia Imporium', bot.user.displayAvatarURL());
+			.setFooter('DMMO Imporium', client.user.displayAvatarURL());
 
 
-		msg.channel.send(embed).then(async sentMessage => {
+		message.channel.send(embed).then(async sentMessage => {
 
 
 			for (let i = 0; i < args.length; i++) {
@@ -39,31 +36,35 @@ module.exports = {
 			}
 
 			item = await profile.getItem(temp);
-			
+
 			if (item) {
-				buy(profile, sentMessage, amount, embed, item, msg);
+				if (item.buyable) buy(profile, sentMessage, amount, embed, item, message);
+				else return sentMessage.edit(embed.setDescription(`${item.name} is not available for purchase.`));
 			}
 			else {
-				msg.channel.awaitMessages(filter, { max: 1, time: 60000 })
+				message.channel.awaitMessages(filter, { max: 1, time: 60000 })
 
 					.then(async collected => {
 						item = await profile.getItem(collected.first().content);
+
 						if (!item) return sentMessage.edit(embed.setDescription(`${collected.first().content} is not a valid item.`));
+						if (!item.buyable) return sentMessage.edit(embed.setDescription(`${item.name} is not available for purchase.`));
+
 						collected.first().delete().catch(e => logger.error(e.stack));
 
 						sentMessage.edit(embed.setDescription(`How many __${item.name}(s)__ do you want to buy?`)).then(() => {
-							msg.channel.awaitMessages(filter, { max: 1, time: 60000 })
+							message.channel.awaitMessages(filter, { max: 1, time: 60000 })
 
 								.then(async collected => {
 									amount = parseInt(collected.first().content);
 									collected.first().delete().catch(e => logger.error(e.stack));
 
-									buy(profile, sentMessage, amount, embed, item, msg);
+									buy(profile, sentMessage, amount, embed, item, message);
 
 								})
 								.catch(e => {
 									logger.error(e.stack);
-									msg.reply('you didn\'t answer in time or something went wrong.');
+									message.reply('you didn\'t answer in time or something went wrong.');
 								});
 						});
 					});
@@ -71,13 +72,12 @@ module.exports = {
 		})
 			.catch(e => {
 				logger.error(e.stack);
-				msg.reply('you didn\'t answer in time or something went wrong.');
+				message.reply('you didn\'t answer in time or something went wrong.');
 			});
-
 	},
 };
 
-async function buy(profile, sentMessage, amount, embed, item, msg) {
+async function buy(profile, sentMessage, amount, embed, item, message) {
 
 	if (!Number.isInteger(amount)) {
 		return sentMessage.edit(embed.setDescription(`**${amount}** is not a number`));
@@ -86,16 +86,15 @@ async function buy(profile, sentMessage, amount, embed, item, msg) {
 		amount = 1;
 	}
 
-	let balance = await profile.getBalance(msg.author.id);
-	const cost = amount * item.cost;
+	let balance = await profile.getBalance(message.author.id);
+	const cost = amount * item.value;
 	if (cost > balance) {
 		return sentMessage.edit(embed.setDescription(`You currently have **${balance}💰**, but __**${amount}**__ __${item.name}(s)__ costs **${cost}💰**!`));
 	}
 
-	await profile.addItem(msg.author.id, item, amount);
-	profile.addMoney(msg.author.id, -cost);
-	profile.addShopSpent(msg.author.id, cost);
+	await profile.addItem(message.author.id, item, amount);
+	profile.addMoney(message.author.id, -cost);
 
-	balance = await profile.getBalance(msg.author.id);
+	balance = await profile.getBalance(message.author.id);
 	sentMessage.edit(embed.setDescription(`You've bought: __**${amount}**__ __${item.name}(s)__.\n\nCurrent balance is **${balance}💰**.`));
 }
